mapping added and new method in user service
